Move shouldNotFailWhenChannelRegistryLevelNull to FetchSerialPublicationByIdentifierAndYearHandlerTest
