+++ src/test/java/no/sikt/nva/pubchannels/handler/fetch/serialpublication/FetchSerialPublicationByIdentifierAndYearHandlerTest.java
@@ -89,6 +89,17 @@ public class FetchSerialPublicationByIdentifierAndYearHandlerTest extends
         return testChannel.asSerialPublicationDto(SELF_URI_BASE, String.valueOf(year));
     }
 
+    @Override
+    protected void mockChannelFound(int year, String identifier, String channelRegistryResponseBody) {
+        mockChannelRegistryResponse(CHANNEL_REGISTRY_PATH_ELEMENT, String.valueOf(year), identifier,
+                                    channelRegistryResponseBody);
+    }
+
+    @Override
+    protected TestChannel generateTestChannel(int year, String identifier) {
+        return new TestChannel(year, identifier, "Series");
+    }
+
     @BeforeEach
     void setup() {
         this.handlerUnderTest = new FetchSerialPublicationByIdentifierAndYearHandler(environment,
@@ -96,23 +107,6 @@ public class FetchSerialPublicationByIdentifierAndYearHandlerTest extends
                                                                                      this.cacheService,
                                                                                      super.getAppConfigWithCacheEnabled(
                                                                                          false));
-    }
-
-    @ParameterizedTest(name = "Should not fail when channel registry level null for type \"{0}\"")
-    @ValueSource(strings = {"Series", "Journal"})
-    void shouldNotFailWhenChannelRegistryLevelNull(String type) throws IOException {
-        var year = randomYear();
-        var identifier = UUID.randomUUID().toString();
-        var testChannel = new TestChannel(year, identifier, type);
-        String channelRegistrySeriesBody = testChannel.asChannelRegistrySeriesBodyWithoutLevel();
-        mockChannelRegistryResponse(CHANNEL_REGISTRY_PATH_ELEMENT, String.valueOf(year), identifier,
-                                    channelRegistrySeriesBody);
-
-        handlerUnderTest.handleRequest(constructRequest(String.valueOf(year), identifier, MediaType.ANY_TYPE), output,
-                                       context);
-
-        var response = GatewayResponse.fromOutputStream(output, SerialPublicationDto.class);
-        assertEquals(HTTP_OK, response.getStatusCode());
     }
 
     @Test
